Divide RPC API tests into mock and remote and use jsonrpc-client
