+++ circuit/src/rpc-test/Cargo.toml
@@ -12,15 +12,22 @@ readme = "README.md"
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
 
-
-
 [dev-dependencies]
-sc-rpc = { version = "3.0.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.8" }
+
+async-std = { version = "1.6.5", features = ["attributes", "tokio1"] }
+async-trait = "0.1.40"
+jsonrpc-runtime-client = { version = "0.1.0", path = "./jsonrpc-runtime-client" }
+jsonrpc-core-client = { version = "15.1.0", features = ["http"] }
+jsonrpsee-types = "0.2.0"
+
+futures = { version = "0.3.16", features = ["compat"] }
+node-primitives = { version = "2.0.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.8" }
 sc-rpc-api = { version = "0.9.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.8" }
+sc-rpc = { version = "3.0.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.8" }
 sc-client-api = { version = "3.0.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.8" }
 sp-api = { version = "3.0.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.8" }
 codec = { package = "parity-scale-codec", version = "2.0.0" }
-futures = { version = "0.3.1", features = ["compat"] }
+
 jsonrpc-pubsub = "15.1.0"
 jsonrpc-core = { version = "15.1.0" }
 hex-literal = "0.3.1"
@@ -61,10 +68,8 @@ pallet-circuit-execution-delivery = { version = "0.1.0", path = "../execution-de
 sc-network = { version = "0.9.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.8" }
 sp-io = { version = "3.0.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.8" }
 substrate-test-runtime = { version = "2.0.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.8" }
-#substrate-test-runtime = { path = "../../../vendor/substrate/test-utils/runtime" }
 substrate-test-runtime-client = { version = "2.0.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.8" }
-#substrate-test-runtime-client = { path = "../../../vendor/substrate/test-utils/runtime/client" }
-bp-runtime = { default-features = false, path = "../../../vendor/bridges/primitives/runtime" }
+bp-runtime = { path = "../../../vendor/bridges/primitives/runtime" }
 
 substrate-frame-rpc-system = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.8" }
 tokio = "0.1.22"
